Make SetContext method atomic.
This allows concurrent goroutines to call `honeybadger.SetContext`. It
does not make `context.Update` safe/atomic because for ease of use,
`Context` must function as a normal map. This means any client code
which calls `context.Update` directly should perform its own locking if
necessary; this may be a case for making `context.Update` a private
method, since it's not really meant to be used directly anyway.