Add information to honeybadger context before calling handler
Because contexts are not modifiable in-place (eg, calling `context.WithValue`
returns a new context, it doesn't return the same context you've passed in with
the value set), need to set some data in the Honeybadger HTTP middleware.

Unfortunately, any additional info set with `SetContext` within the handler is
lost if the handler panics.