Fixes stack traces when methods are inlined
Go introduced runtime.CallerFrames in 1.7 which can expand frames for
inlined methods[1]. They're recommending all code switch to use it.

This really came up because the TestNewErrorTrace was failing on Go
1.12.6 since at some point they added in a line for the method defined
in TestNewErrorTrace. I've expanded the test to make sure the top two
lines are from the code that is calling NewError and doesn't include a
reference to the library itself.

1: https://golang.org/doc/go1.12#compiler